dau修复定时任务，下载图包加入创建文件夹

+++ MarryGuGuNiu.js
@@ -1,3 +1,4 @@
+import fs from 'fs'
 import { exec, execSync } from "child_process"
 
 let url = `https://github.com/GuGuNiu/Miao-Plugin-MBT`
@@ -13,13 +14,8 @@ export class marry_GuGuNiu extends plugin {
             priority: 2955,
             rule: [
                 {
-                    reg: /^#下载咕咕牛图包$/,
+                    reg: /^#(强制)?(更新|下载)咕咕牛图包$/,
                     fnc: 'download_GuGuNiu',
-                    permission: "master"
-                },
-                {
-                    reg: /^#(强制)?更新咕咕牛图包$/,
-                    fnc: 'update_GuGuNiu',
                     permission: "master"
                 },
                 {
@@ -32,31 +28,43 @@ export class marry_GuGuNiu extends plugin {
     }
 
     async download_GuGuNiu(e) {
-        await e.reply(`开始下载${url}`)
-        cmd = `git clone --depth=1 ${url} ${Path}`
-        exec(cmd, { cwd: process.cwd(), stdio: 'inherit' }, (error) => {
-            if (error) return e.reply(`下载错误：\n${error}`)
-            else {e.reply(`下载完成`);this.apply_GuGuNiu()}
-        })
-
-        await e.reply(`下载中，耐心等待，保存路径${Path}`)
-    }
-
-    async update_GuGuNiu(e) {
-        cmd = `git pull`
-        if (this.e.msg.includes('强制'))
-            execSync('git fetch && git reset --hard', { cwd: Path})
-        exec(cmd, { cwd: Path, stdio: 'inherit' }, (output, error) => {
-            if (error) {
-                if(error.match(/Already up to date\./))
-                    e.reply(`咕咕牛正在偷懒`)
-                else
-                    {e.reply(`咕咕牛更新结束`);this.apply_GuGuNiu();}
-            }
-            else {e.reply(`更新错误：${output}`);return}
-        })
-
-        await e.reply(`更新中，耐心等待，保存路径${Path}`)
+        if(!fs.existsSync(`${Path}normal-character` || this.e.msg.includes('下载'))){
+            await this.reply(`开始下载${url}`)
+            cmd = `mkdir ${Path}`
+            if(!fs.existsSync(Path))
+                exec(cmd, { cwd: process.cwd(), stdio: 'inherit' }, (error) => {
+                    if (error) 
+                        this.reply(`文件夹创建错误：\n${error}`)
+                    else 
+                        this.reply(`下载中，耐心等待，保存路径${Path}`)
+                })
+            cmd = `git clone --depth=1 ${url} ${Path}`
+            exec(cmd, { cwd: process.cwd(), stdio: 'inherit' }, (error) => {
+                if (error) 
+                    return this.reply(`下载错误：\n${error}`)
+                else {
+                    this.reply(`下载完成`)
+                    this.apply_GuGuNiu()
+                }
+            })
+        }else{
+            await this.reply(`更新中，耐心等待，保存路径${Path}`)
+            cmd = `git pull`
+            if (this.e.msg.includes('强制'))
+                execSync('git fetch && git reset --hard', { cwd: Path})
+            exec(cmd, { cwd: Path, stdio: 'inherit' }, (output, error) => {
+                if (error) {
+                    if(error.match(/Already up to date\./))
+                        this.reply(`咕咕牛正在偷懒`)
+                    else{
+                        this.reply(`咕咕牛更新结束`)
+                        this.apply_GuGuNiu()
+                    }
+                }
+                else 
+                    return this.reply(`更新错误：${output}`)
+            })
+        }
     }
 
     async apply_GuGuNiu(){
@@ -69,8 +77,10 @@ export class marry_GuGuNiu extends plugin {
             cmd = `cp -a -f -l ${Path}* ./plugins/miao-plugin/resources/profile/`
         
         exec(cmd, { cwd: process.cwd(), stdio: 'inherit' }, (error) => {
-            if (error) return this.reply(`应用错误：\n${error}`)
-            else this.reply(`应用完成`)
+            if (error) 
+                return this.reply(`应用错误：\n${error}`)
+            else 
+                this.reply(`应用完成`)
         })
     }
 }
